Reverted to old mod menu version, was getting loader issues with this one, deleted FabricBannerShieldItem to just have the lone class with components dictating the rest, created api methods
