+++ gradle.properties
@@ -14,6 +14,6 @@ maven_group=com.github.crimsondawn45
 archives_base_name=FabricShieldLib
 
 # Dependencies
-fabric_version=0.124.0+1.21.5
-mod_menu_version=1.16.23
+fabric_version=0.124.2+1.21.5
+mod_menu_version=14.0.0-rc.2
 midnightlib_version=1.7.3+1.21.4-fabric
